Add color code system, cleaned up UI, disabled HUD by default

+++ extract_assets.py
@@ -36,12 +36,13 @@ def asset_needs_update(asset, version):
 
 
 def remove_file(fname):
-    os.remove(fname)
-    print("deleting", fname)
-    try:
-        os.removedirs(os.path.dirname(fname))
-    except OSError:
-        pass
+    if fname != "actors/mario/mario_logo.rgba16.png" and fname != "actors/mario_cap/mario_cap_logo.rgba16.png":
+        os.remove(fname)
+        print("deleting", fname)
+        try:
+            os.removedirs(os.path.dirname(fname))
+        except OSError:
+            pass
 
 
 def clean_assets(local_asset_file):
